TermiteMoundInducedAirflowTrixi.jl API

TermiteMoundInducedAirflowTrixi.TermiteMoundEquations1D — Type

TermiteMoundEquations1D(;γ = 1.3987529976019184, k_i = 0.0004184703194089407, k_w = 4.550553359854458, tᵣ = 87.10175873820496, uᵣ = 0.05, xa = 0.17221939177753745, xb = 0.5680821161575328, xc = 0.9793345160180965, r = 0.6874197366854203, h = 1.5810403656581307, L = 4.355087936910248, β = 26.9749915565812, η = 0.93, Fr² = 5.851592474205325e-5, ρₕ₀ = 0.982, T_ref = 297.76980029566033, t_ref = 0.0033582989242263127, T0 = 27.0, v0 = -2.219, Ti_LI = linear_interpolation([0.0, 0.5, 1.0], [27.6, 28.8, 27.5]))

TermiteMoundEquations1D (TME) in one space dimension. The governing equations are

\[ \begin{aligned} \frac{\partial \rho}{\partial t} + \frac{\partial \left( \rho u\right)}{\partial x} &= - \frac{\textrm{A}_x}{\textrm{A}}\rho u \\ \frac{\partial v}{\partial t} &= \frac{1}{\int_0^1 \frac{\rho}{A} \, dy} \left[ \int_0^1 -\rho u u_x - \rho u \left( \beta \eta - \beta\left(1-\eta\right) \vert u \vert \right) - \frac{\textrm{h}_x}{Fr^2}(\rho - \rho_{h_0}) \, dy \right] \\ \frac{\partial p_0}{\partial t} + \gamma p_0 \frac{\partial u}{\partial x} &= - \gamma \frac{\textrm{A}_x}{\textrm{A}} u p_0 - \frac{k_w}{\textrm{A} \sqrt{\textrm{A}}} \left(T - \textrm{T}_\textrm{u} \right) \\ \frac{\partial T_i}{\partial t} &= k_i \left( T - T_i \right) \\ p_0 &= \rho T \\ u(t,x) &= \frac{v(t)}{A(x)} + \frac{1}{A(x)} \int_0^x A(y)\left[ -\frac{\partial p_0}{\partial t} - \frac{k_w}{\textrm{A} \sqrt{\textrm{A}}} \left(T - \textrm{T}_\textrm{u} \right)\right] \, dy. \end{aligned}\]

The unknown quantities of the TME are the air density $ρ$, the air velocity $u$, the leading order pressure $p_0$, the internal temperature of the mound $T_i$, and the auxiliary velocity variable $v$. We denote the (possibly) variable cross section function $A(x)$, the height profile function $h(x)$ and the boundary temperature function $T_u(t,x)$. The boundary temperature function is defined piecewise by

\[ \begin{align*} \textrm{T}_\textrm{u}(t,x) = \begin{cases} \textrm{T}_\textrm{soil}(t) & ,\text{if } x \in [0,x_a] \, \cup \, (x_c, 1] \quad \text{(soil)} \\ \frac{\textrm{T}_\textrm{air}(t) + T_i(t,x) }{2} & ,\text{if } x \in (x_a,x_b] \quad \text{(flute)} \\ T_i(t,x) & ,\text{if } x \in (x_b,x_c] \quad \text{(chimney).} \\ \end{cases} \end{align*}\]

It depends on the internal mound temperature $T_i$, the soil temperature (at 30cm depth) $T_{soil}$ and the ambient air temperature $T_{air}$. The parameters $x_a$, $x_b$ and $x_c$ are fixed spatial locations along the flow channel.

Reference for the TME:

Marx, Oliver P. and Gasser, Ingenuin and Annika Schmidgall (2026) ... DOI: ...

TermiteMoundInducedAirflowTrixi.UpdateVelocityCallback — Type

UpdateVelocityCallback(solver)

Parameters

- `solver`: ODE solver.

The UpdateVelocityCallback called after every time step. In this callback explicit time steps for the velocity $v$ and pressure $p_0$ are performed. For the solver we reccomend CarpenterKennedy2N54 from OrdinaryDiffEqLowStorageRK. After the time integration, the velocity $u$ is computed and the integrator is updated.

Note

If the callback is not used, the quantities $u$, $v$ and $p_0$ will remain constant in time!

TermiteMoundInducedAirflowTrixi.A — Method

A(x, ::TermiteMoundEquations1D)

Defines the cross section of the flow channel.

Parameters

x: Position vector.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Cross section at position x.

TermiteMoundInducedAirflowTrixi.A_x — Method

A(x, ::TermiteMoundEquations1D)

Gradient of the cross section function

Parameters

x: Position vector.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Derivative of the cross section at x.

TermiteMoundInducedAirflowTrixi.Fourir — Method

Fourir(a0,a1,a2,a3,a4,a5,a6,a7,a8,a9,a10,t_var,::TermiteMoundEquations1D,)

Evaluates the function

\[ \begin{align*} F(t) &= a_0 + \sum_{i=1}^5 a_{\left(2i-1\right)} cos\left(\frac{12 \pi t}{i}\right) + a_{\left( 2i\right)} sin\left(\frac{12\pi t}{i}\right) \end{align*}\]

Parameters

a0: Parameter $a_0$.

a1: Parameter $a_1$.

a2: Parameter $a_2$.

a3: Parameter $a_3$.

a4: Parameter $a_4$.

a5: Parameter $a_5$.

a6: Parameter $a_6$.

a7: Parameter $a_7$.

a8: Parameter $a_8$.

a9: Parameter $a_9$.

a10: Parameter $a_{10}$.

t_var: Scalar time.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Function value.

TermiteMoundInducedAirflowTrixi.Get_initial_Ti — Method

Get_initial_Ti(r,h)

Computes an initial profile for the termite mound material temperature.

Parameters

r: Radius.

h: Height.

Returns

Initial mound material temperature profile.

TermiteMoundInducedAirflowTrixi.I_w — Method

I_w(x_var, equations::TermiteMoundEquations1D)

This function defines a variable value of $α_w$ along the spatial domain.

Parameters

x_var: Position variable.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Returns a scalar.

TermiteMoundInducedAirflowTrixi.T_air — Method

T_air(t_var, equations::TermiteMoundEquations1D)

Returns the ambient air temperature $T_{air}$. Parameters are obtained from a fourier series fit of the mean temperature in Figure 2:

Singh, R. K., Sharma, R. V., (2017) Numerical analysis for ground temperature variation DOI: 10.1186/s40517-017-0082-z

Parameters

t_var: Time variable.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Returns ambient air temperature in Kelvin.

TermiteMoundInducedAirflowTrixi.T_soil — Method

T_soil(t_var, equations::TermiteMoundEquations1D)

Returns the soil temperature $T_{soil}$. The parameters are obtained from a fourier series fit of the mean soil temperature at 30 cm depth in Figure 8 and 9 from:

Singh, R. K., Sharma, R. V., (2017) Numerical analysis for ground temperature variation DOI: 10.1186/s40517-017-0082-z

Parameters

t_var: Time variable.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Returns soil temperature in Kelvin.

TermiteMoundInducedAirflowTrixi.T_u — Method

T_u(t_var, x_var, Ti, equations::TermiteMoundEquations1D)

The boundary temperature function is defined piecewise by

\[ \begin{align*} \textrm{T}_\textrm{u}(t,x) = \begin{cases} \textrm{T}_\textrm{soil}(t) & ,\text{if } x \in [0,x_a] \, \cup \, (x_c, 1] \quad \text{(soil)} \\ \frac{\textrm{T}_\textrm{air}(t) + T_i(t,x) }{2} & ,\text{if } x \in (x_a,x_b] \quad \text{(flute)} \\ T_i(t,x) & ,\text{if } x \in (x_b,x_c] \quad \text{(chimney).} \\ \end{cases} \end{align*}\]

It depends on the internal mound temperature $T_i$, the soil temperature $T_{soil}$ (at 30cm depth) and the ambient air temperature $T_{air}$. The parameters $x_a$, $x_b$ and $x_c$ are fixed spatial locations along the flow channel.

Parameters

t_var: Time variable.

x_var: Position variable.

Ti: Inside Material temperature.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Boundary temperature in Kelvin.

TermiteMoundInducedAirflowTrixi.TermiteMoundInitialCondition — Method

TermiteMoundInitialCondition(x, t, equations::TermiteMoundEquations1D)

Defines the initial state of the system.

Parameters

x: Position vector.

t: Time scalar.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Initial Condition as a SVector.

TermiteMoundInducedAirflowTrixi.bspline2linear — Method

bspline2linear(nodes, vals, t, ti, ::TermiteMoundEquations1D)

Creatares an interpolation polynomial for a flow quantity on the entire space domain.

Parameters

nodes: Node vector.

vals: Node values.

t: Uniform nodes (length=Nodes).

ti: Uniform nodes (length>Nodes) .

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Returns an interpolation object of type Interpolations.Extrapolation.

TermiteMoundInducedAirflowTrixi.condition — Method

condition(u, t, integrator)

Parameters

u: State vector.

t: Time scalar.

integrator: ODE integrator.

Ensures that UpdateVelocityCallback is used after every time step.

Returns

True.

TermiteMoundInducedAirflowTrixi.h_x — Method

h_x(x, equations::TermiteMoundEquations1D)

Defines the gradient of the height function of the flow channel.

Parameters

x: Position vector.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Derivative of the height function.

TermiteMoundInducedAirflowTrixi.linear_interpolation2 — Method

linear_interpolation2(x, y, ::TermiteMoundEquations1D)

Modified version of the linear_interpolation function from Interpolations.jl. Identical knots and nodes are identified and removed to prevent warnings.

Parameters

x: Node vector.

y: Function value vector.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Interpolation object of type Interpolations.Extrapolation.

TermiteMoundInducedAirflowTrixi.source_terms — Method

source_terms(u, x, t, equations::TermiteMoundEquations1D)

Defiens the source terms for the TME.

Parameters

u: State vector.

x: Position vector.

t: Time scalar.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Source term as a SVector.

TermiteMoundInducedAirflowTrixi.temp2scaled — Method

temp2scaled(y, equations::TermiteMoundEquations1D)

Converts a scaled temperature into an unscaled temperature in Celsius.

Parameters

y: Scaled temperature.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Unscaled temperature in Celsius.

TermiteMoundInducedAirflowTrixi.temp2unscaled — Method

temp2unscaled(x, equations::TermiteMoundEquations1D)

Converts an unscaled temperature (Celsius) into a scaled temperature.

Parameters

x: Unscaled temperature in Celsius.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Scaled temperature.

TermiteMoundInducedAirflowTrixi.termite_parameters — Method

termite_parameters(r, height; g=9.81, p0=1e5, R=8.314, M=0.0289652, c_v=20.85, α_w = 1.1, D0 = 0.05, ρₕ₀ = 0.982, η = 0.93, uᵣ = 0.05, ρᵣ = 1.17, pᵣ = 1, c_v_ = 790, κ = 0.184, T_ref = 297.76980029566033, t_ref = 0.0033582989242263127, T0 = 27.0, v0 = -5.0)

Initializes the parameters for the termite mound model based on the given radius and height of the mound.

Parameters

r: Position variable.

height: Height of the termite mound.

Optional Parameters

g: Gravitational acceleration (default: 9.81 m/s²).

p0: Reference pressure (default: 1e5 Pa).

R: Universal gas constant (default: 8.314 J/(mol·K)).

M: Molar mass of air (default: 0.0289652 kg/mol).

c_v: Specific heat capacity at constant volume (default: 20.85 J/(mol·K)).

α_w: Heat transfer coefficient (default: 1.1).

D0: Diameter of the flow channel (default: 0.05 m).

ρₕ₀: Reference density (default: 0.982 kg/m³).

η: Viscosity (default: 0.93).

uᵣ: Reference velocity (default: 0.05 m/s).

ρᵣ: Reference density (default: 1.17 kg/m³).

pᵣ: Reference pressure (default: 1 Pa).

c_v_: Specific heat capacity of the mound material (default: 790 J/(kg·K)).

κ: Thermal conductivity of the mound material (default: 0.184 W/(m·K)).

T_ref: Reference temperature (default: 297.76980029566033 K).

t_ref: Reference time (default: 0.0033582989242263127 s).

T0: Initial temperature (default: 27.0 °C).

v0: Initial velocity (default: -5.0 m/s).

Returns

A 20-tuple which can be used to create an instance of TermiteMoundEquations1D: (γ, k_i, k_w, tᵣ, uᵣ, xa, xb, xc, r, h, L, β, η, Fr², ρₕ₀, T_ref, t_ref, T0, v0, Ti_LI).

TermiteMoundInducedAirflowTrixi.unique_idx — Method

unique_idx(x, ::TermiteMoundEquations1D)

This function identifies a set of unique spatial positions given all elements of a discretisation.

Parameters

x: Position variable.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Returns a node vector with unique elements and their corresponding indices.

TermiteMoundInducedAirflowTrixi.vel2scaled — Method

vel2scaled(y, equations::TermiteMoundEquations1D)

Converts a scaled velocity into unscaled velocity in cm/s.

Parameters

y: Scaled velocity.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Unscaled velocity in cm/s.

TermiteMoundInducedAirflowTrixi.vel2unscaled — Method

vel2unscaled(x, equations::TermiteMoundEquations1D)

Converts an unscaled velocity (cm/s) into a scaled velocity.

Parameters

x: Unscaled velocity.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Scaled velocity.

Trixi.flux — Method

Trixi.flux(u, orientation::Integer, ::TermiteMoundEquations1D)

Calculate 1D flux for a single point

Parameters

u: State vector.

orientation: Orientation or normal direction.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Flux as a SVector.

Trixi.flux_ranocha — Method

Trixi.flux(u, orientation::Integer, ::TermiteMoundEquations1D)

Calculate 1D flux.

Parameters

u: State vector.

orientation: Orientation or normal direction.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Flux as a SVector.

Trixi.max_abs_speeds — Method

Trixi.max_abs_speeds(u, equations::TermiteMoundEquations1D)

Computes the maximum absolute speed for wave propagation.

Parameters

u: State vector.

equations::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

Maximum absolute speed.

Trixi.varnames — Method

Trixi.varnames(u, ::TermiteMoundEquations1D)

Returns the variable names corresponding to the conserved variables in the termite mound model.

Parameters

::typeof(cons2cons): Type indicating conserved to conserved variable conversion.

::TermiteMoundEquations1D: Instance of TermiteMoundEquations1D.

Returns

A tuple of variable names: ("rho", "v1", "p0", "Ti", "x_var").